Move side config setters into own file
